+++ CHANGELOG.md
@@ -22,6 +22,10 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * Added a new static method `wrapReasonToIgnore` `core/promise/abortable`
 * Added static properties with the default error types `core/request/error`
 
+#### :bug: Bug Fix
+
+* Fixed a bug when a child promise never resolved because its parent is already rejected `core/promise/abortable`
+
 ## v3.68.0 (2021-12-03)
 
 #### :boom: Breaking Change